verify email by otp  , signup , logout , login files are working fine and tested

+++ backend/utils/generateAndStoreOtp.js
@@ -3,6 +3,7 @@ import { configDotenv } from 'dotenv';
 import client from '../libs/redis.js';
 import generateCooldownTime from './generateCooldownTime.js';
 import { raw } from 'express';
+import User from '../models/user.model.js';
 configDotenv() ;
 
 export const hashOTP = async (otp  , email) => {
@@ -14,11 +15,10 @@ export const hashOTP = async (otp  , email) => {
 
 
 
-export const generateVerificationCode = () =>{
-
-
-
-    const generateOTP = async() => {
+export const generateVerificationCode = async  (email) =>{
+       
+     try {
+        const generateOTP = async() => {
         const n = crypto.randomInt(0 , 1000000) ;
         console.log(n) ;
         return String(n).padStart(6 , "0");
@@ -29,8 +29,9 @@ export const generateVerificationCode = () =>{
     const StoreHashOtpRedis = async (email) => {
 
         const key = `otp:${String(email).toLocaleLowerCase()}` ;
-        const otp = generateOTP() ;
-        const hash = hashOTP(otp , email) ;
+        const otp = await generateOTP() ;
+        const hash = await hashOTP(otp , email) ;
+
 
         const payload = JSON.stringify({
             hash , 
@@ -44,54 +45,64 @@ export const generateVerificationCode = () =>{
        
     }
 
-    //   generateOTP() ;
-    //   hashOTP() ;
-      return  StoreHashOtpRedis() ;
+       return  StoreHashOtpRedis(email) ;
+     } catch (error) {
+         console.log("error in generating the otp for email verification ", error.message) ;
+     }
+    
     
 }
 
 export const verifyOtp = async (req , res) => {
      try {
         const providedOtp = String(req.body.otp ?? "") ;
-        const sid = req.cookies ?.otp_sid || req.query.sid ; //req.query.sid => from url for cross-device
+        const sid = req.cookies.otp_sid || req.query.sid ; //req.query.sid => from url for cross-device
+
+        
 
         if(!sid){
-             res.status(440).json({error : "otp session is expired"}) ;
+            return res.status(440).json({error : "otp session is expired in browser"}) ;
         }
 
-        const rawEmail = await client.get(`otpSessionId:${sid}`) ;
+        const email = await client.get(`otpSessionId:${sid}`) ;
 
-        if(!rawEmail){
-            res.status(400).json({error:"the otp session is expireed in redis"}) ;
+        if(!email){
+            return res.status(400).json({error:"the otp session is expireed in redis"}) ;
         }
 
-        const email = JSON.parse(rawEmail) ;
+        // const email = await JSON.parse(rawEmail) ;
         
         //cool-down period 
         if(await client.get(`coolDownOtp:${email}`)){
-            res.status(429).json({error : "cooldown_active"}) ;
+            await client.del(`otp:${email}`) ;
+            return res.status(429).json({error : "cooldown_active"}) ;
         }
         //check otp 
-        const rawOtp = client.get(`otp:${email}`) ;
-        const otp_Meta = JSON.parse(rawOtp);
-        const hash = await crypto.createHmac("256" , process.env.HMAC_OTP_SECRET).update(email + '|' + providedOtp).digest("hex");
+        const rawOtp = await client.get(`otp:${email}`) ;
+        const otp_Meta = await JSON.parse(rawOtp);
+        const hash = await hashOTP(providedOtp , email)
+        
         
         if(hash != otp_Meta.hash){
             otp_Meta.attempts_left -= 1 ;
             if(otp_Meta.attempts_left <= 0){
                 await client.del(`otp:${email}`) ;
-                generateCooldownTime() ;
+                await generateCooldownTime(email) ;
                 return res.status(403).json({error:"the otp is incorrect wait for 4 hours"}) ;
             }
             await client.set(`otp:${email}` , JSON.stringify(otp_Meta) , "EX" , await client.ttl(`otp:${email}`)) ;
-            return req.status(401.).json({error:"invalid otp" , attemptsLeft : otp_Meta.attempts_left}) ;
+            return res.status(401.).json({error:"invalid otp" , attemptsLeft : otp_Meta.attempts_left}) ;
         }
 
         await client.del(`otp:${email}`) ;
         await client.del(`otpSessionId:${sid}`) ;
         await client.del(`coolDownOtp:${email}`) ;
+        await User.updateOne(
+            {email : email} ,
+            {$set : {isVerified: true}}
+        ) ;
         res.clearCookie("otp_sid") ;
-        res.status(200).json({success : true , message : "user verified"}) ;
+        return res.status(200).json({success : true , message : "user verified"}) ;
